make it possible to have more than 4B vertices
change vid_t from u32 to u64 in order to handle very big graph

+++ src/graphchi_types.hpp
@@ -23,7 +23,7 @@
 
 namespace graphchi {
     
-    typedef uint32_t vid_t;
+    typedef uint64_t vid_t;
     
     
     /**
